requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3

# LangChain
langchain==0.3.13
langchain-groq==0.2.4
langchain-community==0.3.13

# Embeddings & Vector Store
sentence-transformers==3.3.1
pinecone-client==5.0.1

# Data generation
faker==33.1.0

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
